Rename Persister::ArHelper to Core::ArHelper
