+++ bin/console
@@ -1,8 +1,8 @@
 #!/usr/bin/env ruby
 
 require "bundler/setup"
-require "topological_inventory/persister/ar_helper"
-TopologicalInventory::Persister::ArHelper.load_environment!
+require "topological_inventory/core/ar_helper"
+TopologicalInventory::Core::ArHelper.load_environment!
 
 # You can add fixtures and/or initialization code here to make experimenting
 # with your gem easier. You can also use a different console, if you like.
